backend/pyproject.toml
[project]
name = "evolution-of-todo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "bcrypt>=4.0.0,<5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-genai>=1.61.0",
    "groq>=1.0.0",
    "mcp>=1.26.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
